+++ Meta/Lagom/Tools/CodeGenerators/LibWeb/CMakeLists.txt
@@ -4,5 +4,5 @@ lagom_tool(Generate_CSS_PropertyID_h   SOURCES Generate_CSS_PropertyID_h.cpp)
 lagom_tool(Generate_CSS_PropertyID_cpp SOURCES Generate_CSS_PropertyID_cpp.cpp)
 lagom_tool(Generate_CSS_ValueID_h      SOURCES Generate_CSS_ValueID_h.cpp)
 lagom_tool(Generate_CSS_ValueID_cpp    SOURCES Generate_CSS_ValueID_cpp.cpp)
-lagom_tool(WrapperGenerator            SOURCES WrapperGenerator.cpp)
-target_compile_options(WrapperGenerator PUBLIC -g)
+
+add_subdirectory(WrapperGenerator)
