+++ Meta/Lagom/Tools/CodeGenerators/LibWeb/WrapperGenerator/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(SOURCES "")
+
+lagom_tool(WrapperGenerator SOURCES
+    IDLGenerators.cpp
+    IDLParser.cpp
+    main.cpp)
+
+target_compile_options(WrapperGenerator PUBLIC -g)
